+++ order-service/src/main/java/com/store/order/application/service/OrderService.java
@@ -1,12 +1,14 @@
 package com.store.order.application.service;
 
+import com.store.order.domain.exceptions.OrderCreationException;
+import com.store.order.domain.exceptions.OrderNotFoundException;
+import com.store.order.domain.exceptions.OrderUpdateException;
 import com.store.order.domain.model.Order;
 import com.store.order.domain.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class OrderService {
@@ -16,12 +18,17 @@ public class OrderService {
     
     
     public Order createOrder(Order order) {
-        order.setStatus("CREATED");
-        return orderRepository.save(order);
+        try {
+            order.setStatus("CREATED");
+            return orderRepository.save(order);
+        } catch (Exception e) {
+            throw new OrderCreationException(e.getMessage());
+        }
     }
 
-    public Optional<Order> getOrderById(Long id) {
-        return orderRepository.findById(id);
+    public Order getOrderById(Long id) {
+        return orderRepository.findById(id)
+                .orElseThrow(() -> new OrderNotFoundException(id));
     }
 
     public List<Order> getOrders() {
@@ -29,12 +36,21 @@ public class OrderService {
     }
 
     public Order updateOrderStatus(Long orderId, String status) {
-        Order order = orderRepository.findById(orderId).orElseThrow();
-        order.setStatus(status);
-        return orderRepository.save(order);
+        Order order = orderRepository.findById(orderId)
+                .orElseThrow(() -> new OrderNotFoundException(orderId));
+
+        try {
+            order.setStatus(status);
+            return orderRepository.save(order);
+        } catch (Exception e) {
+            throw new OrderUpdateException(orderId, status);
+        }
     }
 
     public void deleteOrderById(Long id) {
+        if (!orderRepository.existsById(id)) {
+            throw new OrderNotFoundException(id);
+        }
         orderRepository.deleteById(id);
     }
 }
